Add level name for stage RTAs in the queue list. (#5)
This pull requests adds the level for which a Stage RTA was submitted in
the queue list.

Previously, line would look like this

> Stage RTA in 5:00 by runner

After this change, the line will look like this:

> Stage RTA (Whomps Fortress) in 5:00 by runner

The changes fetch the level data from the game info (refreshed once a
day) and the level from each run (null when unset), parses that
information out and displays it. The row of runs has been allowed to
expand 200px more to help keep all the information on one line on screen
that can allow it.